requirements.txt
# API clients
anthropic>=0.8.0
openai>=1.3.0
elevenlabs>=0.2.24

# Utilities
python-dotenv>=1.0.0
moviepy>=1.0.3
pydub>=0.25.1
